docs: introduce docs site
