+++ docs/overrides/main.html
@@ -1,0 +1,7 @@
+{% extends "base.html" %}
+
+{#
+{% block announce %}
+  <p>🎉 Docling is now officially supported in LlamaIndex! <a href="{{ 'integrations/llamaindex/' | url }}">Check it out</a>!</p>
+{% endblock %}
+#}
