+++ docs/stylesheets/extra.css
@@ -1,0 +1,3 @@
+[data-md-color-scheme="default"] .md-banner a {
+    color: #5e8bde;
+}
