+++ pyproject.toml
@@ -72,6 +72,8 @@ pandas-stubs = "^2.1.4.231227"
 ipykernel = "^6.29.5"
 ipywidgets = "^8.1.5"
 nbqa = "^1.9.0"
+mkdocs-material = "^9.5.40"
+mkdocs-jupyter = "^0.25.0"
 
 [tool.poetry.group.examples.dependencies]
 datasets = "^2.21.0"
